Fix web and mobile sidebar navigation
