+++ packages/commonwealth/client/scripts/views/components/component_kit/cw_mobile_menu.tsx
@@ -24,7 +24,7 @@ const CWMobileMenuItem = (props: MenuItem) => {
         onClick={(e) => {
           // Graham TODO 22.10.06: Temporary solution as we transition Notifications
           app.mobileMenu = null;
-          app.mobileMenuRedraw.emit('redraw');
+          app.sidebarRedraw.emit('redraw');
           onClick(e);
         }}
       >
@@ -50,7 +50,7 @@ const CWMobileMenuItem = (props: MenuItem) => {
         onClick={(e) => {
           // Graham TODO 22.10.06: Temporary solution as we transition Notifications
           app.mobileMenu = null;
-          app.mobileMenuRedraw.emit('redraw');
+          app.sidebarRedraw.emit('redraw');
           onClick(e);
         }}
       >
