+++ packages/commonwealth/client/scripts/views/components/sidebar/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useLocation } from 'react-router-dom';
 
 import 'components/sidebar/index.scss';
@@ -15,7 +15,6 @@ import { SidebarQuickSwitcher } from './sidebar_quick_switcher';
 import { CWIcon } from '../component_kit/cw_icons/cw_icon';
 import { CWText } from '../component_kit/cw_text';
 import { useCommonNavigate } from '../../../navigation/helpers';
-import useForceRerender from 'hooks/useForceRerender';
 
 export type SidebarMenuName =
   | 'default'
@@ -24,16 +23,7 @@ export type SidebarMenuName =
 
 export const Sidebar = () => {
   const navigate = useCommonNavigate();
-  const forceRerender = useForceRerender();
   const { pathname } = useLocation();
-
-  useEffect(() => {
-    app.sidebarRedraw.on('redraw', () => forceRerender());
-
-    return () => {
-      app.sidebarRedraw.off('redraw', () => forceRerender());
-    };
-  });
 
   const onHomeRoute = pathname === `/${app.activeChainId()}/feed`;
